+++ app/routes/index.js
@@ -13,7 +13,10 @@ export default class IndexRoute extends Route {
   beforeModel() {
     // we are logged in
     // => go to dashboard.home
-    if (this.session.isAuthenticated) {
+    if (
+      this.session.isAuthenticated &&
+      this.session.data.authenticated.kind !== 'payment-approval' // don't redirect to dashboard if we logged in via payment-approval key
+    ) {
       this.router.transitionTo(NATIVE_MOBILE_ROUTE.DASHBOARD.HOME);
       return;
     }