+++ CHANGELOG.md
@@ -1,3 +1,95 @@
+## [1.18.10](https://github.com/houseninjadojo/app/compare/v1.18.9...v1.18.10) (2022-07-17)
+
+
+### Bug Fixes
+
+* **deep-links:** just use raw url ([#791](https://github.com/houseninjadojo/app/issues/791)) ([155df80](https://github.com/houseninjadojo/app/commit/155df80a2fa7d017a7174bde89967732317ec152))
+
+## [1.18.9](https://github.com/houseninjadojo/app/compare/v1.18.8...v1.18.9) (2022-07-17)
+
+
+### Bug Fixes
+
+* **branch:** only use test key in simulator ([#790](https://github.com/houseninjadojo/app/issues/790)) ([c2b45fe](https://github.com/houseninjadojo/app/commit/c2b45fe98bcd2f9ccd8b3309a759cdfd48eca114))
+
+## [1.18.8](https://github.com/houseninjadojo/app/compare/v1.18.7...v1.18.8) (2022-07-17)
+
+
+### Bug Fixes
+
+* **deep-links:** fix login ([#789](https://github.com/houseninjadojo/app/issues/789)) ([be9376d](https://github.com/houseninjadojo/app/commit/be9376dfa2b5ac9cf178a3c8649c69dfd04c2ea1))
+
+## [1.18.7](https://github.com/houseninjadojo/app/compare/v1.18.6...v1.18.7) (2022-07-16)
+
+
+### Bug Fixes
+
+* deep links ([#788](https://github.com/houseninjadojo/app/issues/788)) ([2fa71a9](https://github.com/houseninjadojo/app/commit/2fa71a97b9a3605b81cd9a62cd0cdffb841c3018))
+
+## [1.18.6](https://github.com/houseninjadojo/app/compare/v1.18.5...v1.18.6) (2022-07-16)
+
+
+### Bug Fixes
+
+* **deep-links:** branch link deep linking ([#787](https://github.com/houseninjadojo/app/issues/787)) ([22a0e5c](https://github.com/houseninjadojo/app/commit/22a0e5cdce656fdb0a9973618e3cd9cf36595d03))
+
+## [1.18.5](https://github.com/houseninjadojo/app/compare/v1.18.4...v1.18.5) (2022-07-16)
+
+
+### Bug Fixes
+
+* **settings:** fix payment method save action ([#786](https://github.com/houseninjadojo/app/issues/786)) ([cb429e9](https://github.com/houseninjadojo/app/commit/cb429e97bceac739be9ff824c34957f5febec756))
+
+## [1.18.4](https://github.com/houseninjadojo/app/compare/v1.18.3...v1.18.4) (2022-07-15)
+
+
+### Bug Fixes
+
+* post-password setting message when setting up account ([#785](https://github.com/houseninjadojo/app/issues/785)) ([6241d95](https://github.com/houseninjadojo/app/commit/6241d95f6b0ceb8a0cdec322d96602780a887232))
+
+## [1.18.3](https://github.com/houseninjadojo/app/compare/v1.18.2...v1.18.3) (2022-07-15)
+
+
+### Bug Fixes
+
+* post-external-payment-approval-response dialog ([#784](https://github.com/houseninjadojo/app/issues/784)) ([950dcc3](https://github.com/houseninjadojo/app/commit/950dcc38d0b618d61d189f213de99f1a31e4089f))
+
+## [1.18.2](https://github.com/houseninjadojo/app/compare/v1.18.1...v1.18.2) (2022-07-15)
+
+
+### Bug Fixes
+
+* **invoice:** mark invoice notes htmlSafe ([#783](https://github.com/houseninjadojo/app/issues/783)) ([a06a841](https://github.com/houseninjadojo/app/commit/a06a84137b69ced1d2ee117635ecabc19fc9c0bf))
+
+## [1.18.1](https://github.com/houseninjadojo/app/compare/v1.18.0...v1.18.1) (2022-07-15)
+
+
+### Bug Fixes
+
+* **invoice:** format description/notes with line breaks ([#782](https://github.com/houseninjadojo/app/issues/782)) ([b7a1436](https://github.com/houseninjadojo/app/commit/b7a143671b06e78ff93a46e92d9c5a50e5338b51))
+
+# [1.18.0](https://github.com/houseninjadojo/app/compare/v1.17.8...v1.18.0) (2022-07-15)
+
+
+### Bug Fixes
+
+* **auth:** check auth kind in index route ([#780](https://github.com/houseninjadojo/app/issues/780)) ([604a9d9](https://github.com/houseninjadojo/app/commit/604a9d983146ff557422e78dbb0798ce10ea9978))
+* **invoice:** use receipt over invoice ([#777](https://github.com/houseninjadojo/app/issues/777)) ([4a273b6](https://github.com/houseninjadojo/app/commit/4a273b6c9584c6449f798440aa57cfa989bfb7ae))
+
+
+### Features
+
+* expire local storage items after some time ([#754](https://github.com/houseninjadojo/app/issues/754)) ([d67991f](https://github.com/houseninjadojo/app/commit/d67991f5ae2bfbfd1779c42944333c675347d294))
+
+## [1.17.8](https://github.com/houseninjadojo/app/compare/v1.17.7...v1.17.8) (2022-07-15)
+
+
+### Bug Fixes
+
+* move payment-approval redirect logic to session service ([#779](https://github.com/houseninjadojo/app/issues/779)) ([5505045](https://github.com/houseninjadojo/app/commit/55050456695ec75c405e68092d77175582832e15))
+
+## [1.17.7](https://github.com/houseninjadojo/app/compare/v1.17.6...v1.17.7) (2022-07-14)
+
 ## [1.17.6](https://github.com/houseninjadojo/app/compare/v1.17.5...v1.17.6) (2022-07-14)